[gpio: fix signal test]
issue:

The gpio signal test is failing.

reason:

The x86 gpio driver doesn't trigger the interrupt which is needed to
send a signal to user space. The x86 driver starts the interrupt
sequence if a magic value gets written to it. Since it never receives
this value from the tests, it doesn't trigger any interrupts.

Even though the tests attempt to write the magic value, it doesn't turn
up in the x86 driver, since part of the value gets inverted at the gpio
driver due to the device's invert mask.

This problem didn't occur thus far since the device used for the signal
test did use a value of 0 for its invert mask. Hence, the value written
by the test did not get changed. However, with the introduction of the
gpio driver the device configuration is shared between all the x86 gpio
test devices. Since the common configuration has some bits set in the
invert mask, the magic value written by the tests gets altered.

fix:

Update the gpio signal test to consider the device's invert mask.