[term: refactor interface device]
Refactor the terminal driver implementation to match the recent changes
to other interface drivers.

The terminal driver implementation already provided an interface driver
between low-level hardware drivers and higher level drivers. However,
instead of being a pure interface driver it was mixed with a devfs
device driver.
In order to match the architecture of the other interface drivers the
terminal driver is split into an interface and a devfs device driver.
Further, the terminal driver helper functions are reorganised to only
expose the necessary functions to the rest of the kernel.