[devtree scripts: unroll arch and memory scripts]
Unroll the arch and memory scripts into the base script rather than
including them.

arch and memory devtree scripts are rather unique and unlikely to be
shared among base scripts. Further, it is a little annoying having to
open multiple files when wanting to work on a targets memory or arch
setup.