[devtree scripts: unroll device scripts]
Unroll the device scripts into the base script rather than including
them.

device devtree scripts are specific to each base script and hence
unlikely to be shared among them. As for arch and memory devtree scripts
it is a little annoying having to open multiple files when wanting to
work on a targets devices.