[errno: fix return value in return_errno]
issue:

return_errno doesn't return a negative value even though it should.

reason:

The macro parameter to return_errno is braced incorrectly when creating
the return value. Instead of having the dash outside of the brackets, it
is inside.

In the case that a non-trivial value, e.g. by using the ternary operator,
is passed to return_errno, the dash doesn't negate the entire expression.
If an expression such as "errno ? errno : E_INVAL" is passed and errno is
not 0, E_INVAL instead of -E_INVAL is returned.

fix:

Add an intermediate variable for the macro parameter. With that the
bracing of the return value is not needed and potential side effects of
the parameter are only triggered once.