[dtc: merge keywords baseaddr and regs]
Merge the baseaddr and reg keywords into the new keyword addr, keeping
it a single pointer rather than a pointer and a list of pointers.

Both keywords represent a pointer with the difference that baseaddr is a
single pointer while reg expects a list of pointers. With the recent
introduction of device macros, having support for a list of pointers
doesn't provide too much of a benefit since repetition of the statement
is only required within the device macros. Hence, it is deemed more
beneficial to reduce the complexity of the dtc parser.