[dtc: revise integer syntax]
Replace the arbitrary-width integer syntax with fixed-with integers and
remove integer lists.

While the arbitrary-width integer syntax allows to freely choose the
width of an integer, it does not represent the use-cases and is also not
supported by the generated c code.

The support for integer lists is dropped in preparation for device
types, which anyhow require to name each member of a type and thus lists
will no longer be needed.