[sys: replace strupr()]
Replace strupr() with its reentrant variante strupr_r().

strupr() used and returned a pointer to a static variable to store the
upper-cased string. Firstly, this needs static memory. Secondly, if
strupr() is called multiple times in the same callstack the result of an
earlier call might not be as expected, depending on the call order.

Replacing strupr() saves the memory and prevent the confusion.